feat: add functionality to run plugin in either environment or pre-command hook (#27)
Adds functionality to run the plugin in either the environment or
pre-command hook. This is useful when using this plugin with other
plugins that may share environment variables used for authentication.
For example, the artifacts plugin uses the same gcloud default application
credentials environment variable to authenticate to GCS and it might be
preferable to have the artifacts plugin run before this plugin. Thus,
allowing the artifacts plugin to use credentials set for the node and
the command hook to use oidc credentials set with this plugin.

I also sorted the plugin configuration properties in the README and plugin.yml

Fixes #26